Merges InfiniteJobs into JobsList component
